refactor: modulise CreatBillForm
