+++ src/pages/BillsPage.js
@@ -1,40 +1,26 @@
 import React, { useContext, useEffect, useState, useRef } from "react";
 import { AuthContext } from "components/auth/Auth";
-import { Formik, Form } from "formik";
 import { Alert } from "react-bootstrap";
-import { isNumeric } from "validator";
 import useSWR from "swr";
 
 import Layout from "components/layout/Layout";
-import SubmitButtonSpinner from "components/forms/SubmitSpinnerButton";
 import Card from "components/Card";
 import TransactionTable from "components/transactions/TransactionTable";
 import gpib from "apis/gpib";
-import Input from "components/forms/Input";
 import BillsHistoryTable from "components/bills/BillsHistoryTable";
-import ToggleButton from "components/forms/ToggleButton";
 import PayWithCustodialWalletModal from "components/bills/PayWithCustodialWalletModal";
 import PayWithPersonalWalletModal from "components/bills/PayWithPersonalWalletModal";
-import ErrorMessage from "components/ErrorMessage";
+import CreateBillForm from "components/bills/CreateBillForm";
 import "./Dashboard.scss";
-
-const validate = ({ billercode, reference, fiat }) => {
-  const errors = {};
-  const reqMsg = "This field is required";
-  if (!billercode) errors.billercode = reqMsg;
-  if (!reference) errors.ref = reqMsg;
-  if (!isNumeric(String(fiat))) errors.fiat = "Amount must be a number";
-  return errors;
-};
 
 const BillsPage = () => {
   const { user } = useContext(AuthContext);
   // form
   const [showModal, setShowModal] = useState(false);
-  const [custodialAddressMessage, setCustodialAddressMessage] = useState("");
+  const [hasCustodialAddress, setHasCustodialAddress] = useState(false);
   const [isPaid, setIsPaid] = useState(false);
   const [paymentAddress, setPaymentAddress] = useState("");
-  const [payWithGpibCustodialWallet, setPayWithGpibCustodialWallet] =
+  const [payWithCustodialWallet, setPayWithCustodialWallet] =
     useState(false);
   const [billInstructions, setBillInstructions] = useState(
     "Fetching your unique payment address... (NOT IMPLEMENTED)"
@@ -42,7 +28,6 @@ const BillsPage = () => {
   const [billBtcAmount, setBillBtcAmount] = useState(0.0008888);
   const [errorMessage, setErrorMessage] = useState(null);
   const [custodialBtcBalance, setCustodialBtcBalance] = useState(null);
-
   const [billId, setBillId] = useState(null);
 
   const { data: bills, error: fetchBillsError } = useSWR(`/bills`);
@@ -69,10 +54,10 @@ const BillsPage = () => {
       const custodialAddress = await getCustodialAddress(addresses);
       const btcBalances = await getUserAddressBalances();
       if (!custodialAddress || !btcBalances) {
-        setCustodialAddressMessage("(No active custodial wallet found)");
         return;
       }
-
+      
+      setHasCustodialAddress(true);
       setCustodialBtcBalance(btcBalances[custodialAddress]);
     };
     fetchData();
@@ -144,13 +129,6 @@ const BillsPage = () => {
     }
   };
 
-  const initialValues = {
-    label: "",
-    billercode: "",
-    reference: "",
-    fiat: 0
-  };
-
   return (
     <Layout activeTab="bills">
       <div className="bills container py-5">
@@ -163,38 +141,14 @@ const BillsPage = () => {
             BPay Biller code and Reference number from your bill and we will pay
             it instantly.
           </Alert>
-          <Formik
-            initialValues={initialValues}
-            validate={validate}
+          <CreateBillForm
             onSubmit={onSubmit}
-          >
-            <Form>
-              <Input
-                name="label"
-                label="Label for your reference"
-                placeholder="Rent, Power Bill, etc"
-              />
-              <Input name="billercode" label="Biller Code" />
-              <Input name="reference" label="Ref" />
-              <Input name="fiat" label="AUD Amount" />
-              <div>
-                <ToggleButton
-                  name="payWithGpibCustodialWallet"
-                  label={`Pay with GPIB custodial wallet ${custodialAddressMessage}`}
-                  onClick={(e) => {
-                    setPayWithGpibCustodialWallet(e.target.checked);
-                  }}
-                  disabled={custodialBtcBalance === null}
-                />
-              </div>
-              <SubmitButtonSpinner
-                isSubmitting={showModal}
-                className="mt-3"
-                submitText="Pay now with Bitcoin"
-              />
-              <ErrorMessage error={errorMessage} isHidden={!errorMessage || showModal} />
-            </Form>
-          </Formik>
+            hasCustodialAddress={hasCustodialAddress}
+            setPayWithCustodialWallet={setPayWithCustodialWallet}
+            showModal={showModal}
+            errorMessage={errorMessage}
+            hideError={!errorMessage || showModal}
+          />
         </Card>
 
         <Card>
@@ -203,7 +157,7 @@ const BillsPage = () => {
         </Card>
 
         <PayWithPersonalWalletModal
-          isOpen={showModal && !payWithGpibCustodialWallet}
+          isOpen={showModal && !payWithCustodialWallet}
           isPaid={isPaid}
           paymentAddress={paymentAddress}
           billInstructions={billInstructions}
@@ -212,7 +166,7 @@ const BillsPage = () => {
         />
 
         <PayWithCustodialWalletModal
-          isOpen={showModal && payWithGpibCustodialWallet}
+          isOpen={showModal && payWithCustodialWallet}
           isPaid={isPaid}
           onDismiss={onDismiss}
           custodialBtcBalance={custodialBtcBalance}
